re-arranging the api and moving JSON:API endpoints to a v2 api
