+++ app/application/adapter.js
@@ -2,5 +2,5 @@ import DS from 'ember-data';
 import config from 'client/config/environment';
 
 export default DS.JSONAPIAdapter.extend({
-  host: config.apiURL,
+  host: `${config.apiURL}/v2`,
 });
