+++ app/history/route.js
@@ -3,6 +3,6 @@ import moment from 'moment';
 
 export default Ember.Route.extend({
   model() {
-    return this.store.query('stat', {since: moment().subtract(1, 'days').toDate().getTime()});
+    return this.store.query('state-timebucket', {since: moment().subtract(1, 'days').toDate().getTime()});
   }
 });
